requirements.txt
streamlit
pandas
langchain
groq
python-dotenv
langchain-community
